# fix: performance bottleneck in stat.js