@betamos 2019-10-15
Thanks for the swift response and merge. I didn't realize Stats was load bearing, then perhaps it is wiser to make some of the heavier features configurable at least.

Btw, yesterday it looked like the Windows tests were failing due to this change, but now they look fine. Any idea what happened?