@jacobheun 2019-10-15
> Thanks for the swift response and merge. I didn't realize Stats was load bearing, then perhaps it is wiser to make some of the heavier features configurable at least.

Yeah, Stats definitely needs some love. We're going to be working on the Connection Manager soon and will hopefully work out a minimum set of needed stats (if any) that are required for it to manage connections properly. From there we can shut everything else off by default I think.

> Btw, yesterday it looked like the Windows tests were failing due to this change, but now they look fine. Any idea what happened?

I should have mentioned I reran them. Due to some timing issues on the windows CI environment the tests can be flakey for stats with when they are checked. I'm planning to get that resolved when I go through it as part of the async refactor in the next couple of weeks.